t/100-python.t: Restrict file check to Python files
One of our tests confirms that our 2 Python files are in the repo
directory. Since we're also adding JavaScript files to this directory,
we need to restrict the test to Python files in order to match them
correctly. In this case since we're only testing Python files we don't
really need to worry about the JS files' existence.