Makefile: Simplify `test` target
Remove all the conditions and code copied from Vcsh and only do the bare
minimum necessary to run tests. This will bork in more unexpected ways
if `git` or `prove` aren't present, but it allows us to get rid of the
only GPL-licensed code in the codebase, freeing me up to license the
project in any way I want.